trace triangle heights to surface tiles
