+++ src/scene3d/core/core.pri
@@ -2,10 +2,8 @@ INCLUDEPATH *= $$PWD
 DEPENDPATH *= $$PWD
 
 HEADERS += \
-    $$PWD/draw_utils.h \
     $$PWD/qml_object_names.h \
     $$PWD/scene_object.h
 
 SOURCES += \
-    $$PWD/draw_utils.cpp \
     $$PWD/scene_object.cpp
