+++ src/scene3d/utils/draw_utils.cpp
